+++ src/odoo_yaml_test/__init__.py
@@ -17,7 +17,7 @@ Example:
 from .case import YamlTransactionCase
 from .exceptions import YamlAssertionError, YamlConfigurationError, YamlStepError
 
-__version__ = "0.1.0"
+__version__ = "0.4.0"
 
 __all__ = [
     "YamlAssertionError",